==> notebooks/01_Estruturacao.ipynb <==
# Introdução

Este *notebook* tem como objetivo realizar o primeiro acesso aos dados, identificar o ponto de separação entre os dados que serão usados para treinamento e teste e, por fim, materializar a separação com a criação dos conjuntos de dados.

# Bibliotecas e Funções

# Usado para ler arquivos, carregar código personalizado e ter acesso a outros recursos de sistema
import os
import sys

# Manipulação e análise dos dados
import pandas as pd

# Recursos para visualização dos dados
import matplotlib.pyplot as plt
from IPython.core.display import HTML, display

# Carregar código personalizado disponível em ../src
sys.path.append(os.path.abspath(os.path.pardir))
from src import settings
from src.utils.notebooks import display_side_by_side
from src.utils.experiments import set_dataset_split

# Configurações para a exibição de conteúdo do Pandas e das bibliotecas gráficas
%matplotlib inline 
pd.set_option('display.max_rows', None)
pd.set_option("display.max_columns", None)
pd.set_option('max_colwidth', 150)

# Organização da Estrutura e Download do Conjunto de Dados

Para iniciar o projeto, é preciso criar a estrutura de diretórios de dados, além de baixar e persistir o conjunto de dados no local de referência.

# Criar estrutura de diretórios, caso não exista
if not os.path.exists(settings.DATA_PATH):
    os.makedirs(settings.DATA_PATH)

for directory in ['raw', 'interim', 'processed']:
    path = os.path.join(settings.DATA_PATH, directory)
    if not os.path.exists(path):
        os.makedirs(path)

raw_dataset_path = os.path.join(settings.DATA_PATH, 'raw', 'elo7_recruitment_dataset.csv')

# Baixar e persistir o conjunto de dados, caso ele ainda não esteja em disco
if not os.path.exists(raw_dataset_path):
    raw_dataset = pd.read_csv(settings.RAW_DATASET_PATH)
    assert len(raw_dataset) == 38_507, f'Os dados acessados não têm o número esperado de registros: {len(raw_dataset)} != 38.507).'
    raw_dataset.to_csv(raw_dataset_path, index=False)
    del raw_dataset

# Análise Preliminar

Este é o primeiro ponto de contato com os dados, usado para identificar a quantidade de registros, as colunas disponíveis e os tipos de dados envolvidos. Um ponto importante neste momento é ter o **entendimento inicial do que existe** e **determinar um ponto de corte nos dados para evitar o vazamento de informações** do que será o **conjunto de teste**. 

Separar o conjunto de teste antes mesmo de fazer a análise exploratória completa dos dados é o melhor jeito de simular o que acontecerá com o(s) modelo(s) quando ele for utilizado em produção com dados novos e, com isso, evitar uma estimativa irreal dos resultados.


## Carregamento dos dados

Leitura do arquivo completo disponível.

frame = pd.read_csv(raw_dataset_path, infer_datetime_format=True)

## Visualização dos primeiros registros disponíveis

Nesta primeira visão, é possível entender que o conjunto de dados é contextualizado na sua posição com relação à busca. Mantendo o foco na separação do conjunto de teste, esperaria ter a data da busca para conseguir separar os dados no tempo. A única informação de tempo disponível se refere à criação do produto.

print(f'Total de registros: {len(frame)}')
frame.head(15)

Para confirmar o significado da data com relação ao produto, além de ter a noção de como outras colunas se comportam, é possível fazer a verificação da repetição dos produtos (tabela à esquerda/do topo) e, para o produto mais repetido, verificar as informações disponíveis (tabela à direita/de baixo).

# Criar data frame com o total de ocorrência de cada produto
agg_frame = (
    frame
    [['product_id']]
    .assign(registros=1)
    .groupby('product_id')
    .sum()
    .reset_index()
    .sort_values(by='registros', ascending=False)
)

# Guardar o ID do produto com maior número de ocorrências
most_frequent_product = agg_frame.iloc[0]['product_id']

# Criar dataframe com os registros do produto com maior ocorrência
summary_frame = (
    frame
    .loc[lambda f: f['product_id'] == most_frequent_product]
    .describe(include='all')
    .T
)

# Exibição dos resultados
display_side_by_side([agg_frame.head(20), summary_frame],
                     ['Produtos mais repetidos', 'Resumo Histórico do produto mais repetido'],
                     padding=50)

del agg_frame, summary_frame  # remoção dos dataframes intermediários

Conforme pode ser observado, o produto com maior ocorrência nos dados tem 14 registros. A repetição da data de criação, que é única para todos, o que reforça o entendimento de que se refere à criação do produto e não da busca ou de outro elemento. Os elementos mais próximos de tempo com relação à busca são as colunas de visualizações (*view_counts*) e de pedidos (*order_counts*). Porém, o total de pedidos não tem crescimento monónoto com relação ao total de visualizações. Isto ocorre, de acordo [com a documentação dos dados](https://github.com/SamaraAlves/teste_HT#dataset), porque **os contadores consideram a janela dos últimos 3 meses**.

Um **ponto de alerta**, com relação ao **vazamento de dados**, é que existem duas buscas para o mesmo produto e com total de visualizações diferentes ('presente dia dos pais'). O fato de a busca estar relacionada a uma data do ano reforça a preocupação com aspectos de sazonalidade e diferenças de resultados ao longo do tempo.

O fato de um mesmo produto ter variações de preço e peso é algo curioso. Como o produto tem o mesmo vendedor, é possível entender que houve uma mudança na composição do produto ao longo do tempo e isso gerou a mudança -- e isto não está direcionamente relacionado à quantidade mínima do pedido, já que a mudança do peso não é proporcional à quantidade. A mudança no preço, que também apresentou flutuações, já deve ser esperada em função de promoções e reajustes de preço.

# Filtrar dados para considerar apenas o produto mais frequente e 
# ordenar colunas por visualizações e pedidos
(
    frame
    .loc[lambda f: f['product_id'] == most_frequent_product]
    .sort_values(by=['view_counts', 'order_counts'])
)

del most_frequent_product

## Separação dos dados de Treinamento e Teste

Considerando a visão inicial dos dados e o **entendimento de que a data de criação é a melhor referência para separar os dados entre treinamento e teste**, é preciso analisar como definir o ponto de corte.

A seguir, faz-se a análise do total de produtos disponíveis nos dados caso um determinado período (a combinação de ano-mês) seja utilizado.

def compute_cumulative_products(frame: pd.DataFrame) -> pd.DataFrame:

    agg_frame = (
        frame        
        .assign(period=lambda f: f['creation_date'].apply(lambda x: str(x)[:7]))
        .sort_values(by='period')
        [['period']]
        .assign(produtos=1)
        .groupby('period')
        .sum()
        .reset_index()
        .assign(produtos_totais=lambda f: f['produtos'].cumsum())
        .assign(perc=lambda f: (100 * f['produtos_totais'] / f['produtos'].sum()).apply(lambda x: f'{x:.2f}%'))
    )
    
    return agg_frame

records_frame = compute_cumulative_products(frame).tail(25)
products_frame = compute_cumulative_products(frame.drop_duplicates('product_id')).tail(25)
difference_frame = pd.DataFrame({'Diferença': records_frame['produtos_totais'] - products_frame['produtos_totais']})

display_side_by_side([records_frame, products_frame, difference_frame],
                     ['Registros (A)', 'Produtos Únicos (B)', 'A - B (p/ produtos_totais)'], padding=50)

del records_frame, products_frame, difference_frame

Na tabela à esquerda (*Registros (A)*), é possível observar a totalização de registros, considerando inclusive as repetições de um mesmo produto. Ao centro (*Produtos Únicos (B)*), cada produto é contabilizado de forma única, para que seja possível determinar se a diferença na quantidade de dados será grande para cada ponto de corte. À direita (*A- B (p/ produtos_totais)*) está a diferença absoluta entre os registros acumulados até o período de corte.

Para que se tenha uma quantidade razoável de dados para validar os modelos e experimentos, o ponto de corte definido é para o período de **2018-08**, o que reserva em torno de **15%** dos registros para teste.

Cabe ressaltar que para este conjunto de dados a escolha do ponto de corte representaria uma diferença de 17 meses entre os dados de treino mais recentes e os dados de teste mais recentes. Para um modelo em produção, provavelmente o modelo seria re-treinado algumas vezes nesse período, para acompanhar a mudança do comportamento dos vendedores e seus produtos. Apesar disso, esse período longo de dados de teste reduz (ainda que não evite totalmente) o viés de se ter apenas produtos muito recentes nos dados de teste -- algo que também é amenizado por se ter contadores com uma janela máxima de 3 meses.

Logo abaixo os dados são separados com esse ponto de referência estabelecido.

def compute_split_distribution(frame: pd.DataFrame) -> pd.DataFrame:    
    agg_frame = (
        frame
        [['group']]
        .assign(registros=1)
        .groupby('group')
        .sum()
        .reset_index()
        .assign(perc=lambda f: ( 100 * f['registros'] / f['registros'].sum()).apply(lambda x: f'{x:.2f}%'))
    )

    return agg_frame


cut_off_period = '2018-08'
split_frame = set_dataset_split(frame, cut_off_period)
agg_records = compute_split_distribution(split_frame)
agg_products = compute_split_distribution(split_frame.drop_duplicates('product_id'))

display(HTML(f'<h3>Separação de conjuntos de dados para o período de corte: {cut_off_period}</h3>'))

display_side_by_side([agg_records, agg_products],
                     ['Registros', "Produtos Únicos"],
                     padding=50
                    )
del agg_records, agg_products

Para garantir que a divisão seja válida e não possua alguma limitação com relação às classes disponíveis, duas validações:
 - Não há períodos em comum entre os dois conjuntos;
 - As categorias do conjunto de teste também existem no conjunto de treinamento.

## Persistência dos Dados de Treinamento e de Teste

Logo abaixo, os dois conjuntos de dados são verificados com relação à separação de períodos e produtos.

training_frame = split_frame.loc[lambda f: f['group'] != 'test'].drop(columns=['group'])
test_frame = split_frame.loc[lambda f: f['group'] == 'test'].drop(columns=['group'])
del split_frame

assert len(set(test_frame['category'].tolist()) - set(training_frame['category'].tolist()) ) == 0, \
'Há categorias do grupo de teste que não estão no grupo de treinamento.'

assert len(set(test_frame['period'].tolist()) & set(training_frame['period'].tolist()) ) == 0, \
'Há períodos compartilhados entre os dados de treinamento e de teste.'

Considernado que as premissas básicas não foram violadas, os conjuntos de dados serão persistidos para uso nas outras etapas.

training_frame.to_csv(os.path.join(settings.DATA_PATH, 'interim', 'training.csv'), index=False)
test_frame.to_csv(os.path.join(settings.DATA_PATH, 'interim', 'test.csv'), index=False)
