==> notebooks/04.4_Historico_e_Analise_de_Experimentos.ipynb <==
# Histórico e Análise de Experimentos de Classificação de Intenções de Busca

Este *notebook* tem como objetivo permitir o registro visual dos experimentos feitos e oferecer um modo de recuperar um experimento e analisar os resultados em mais detalhes.

## Bibliotecas e Funções

# General
import sys
import funcy as fp
from pathlib import Path

# Visualization / Presentation
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.core.display import HTML, display

import mlflow
from mlflow.tracking import MlflowClient
import pandas as pd

# Carregar, além de atualizar frequentemente, código personalizado disponível em ../src
%load_ext autoreload 
%autoreload 2
sys.path.append(str(Path.cwd().parent))
from src import settings
from src.utils.notebooks import display_side_by_side
from src.pipeline.inference_pipeline import load_model_resources

# Configurações para a exibição de conteúdo do Pandas e das bibliotecas gráficas
%matplotlib inline 
sns.set(rc={'figure.figsize':(25,10)})
pd.set_option('display.max_rows', None)
pd.set_option("display.max_columns", None)
pd.set_option('max_colwidth', 150)

## Recuperação do melhor resultado

Considerando os experimentos feitos no notebook [Classificação de Intenções](04.3_Classificacao_de_Intencoes.ipynb), faz-se a recuperação dos melhores resultados de cada algoritmo, tendo como métrica principal o F1.

EXPERIMENT_ID = '2'

mlflow_client = MlflowClient()

best_experiments_result = [
    mlflow.search_runs(experiment_ids=[experiment_id], 
                       max_results=100, 
                       order_by=['metrics.F1 DESC'], 
                       filter_string='attributes.status="FINISHED"')
    for experiment_id in [EXPERIMENT_ID]
]

best_results = pd.concat(best_experiments_result, axis=0)

O melhores resultados recuperados:

columns_to_show = ['experiment_name', 'tags.mlflow.runName', 'run_id', 'experiment_id', 'params.model_name',
                   'metrics.f1', 'metrics.precision', 'metrics.recall', 'metrics.auc', 'metrics.mc', 'metrics.training_time']

(best_results
 .assign(experiment_name=lambda f: f['experiment_id'].apply(lambda id: mlflow_client.get_experiment(id).name))
 .sort_values(by='metrics.f1', ascending=False)
 .loc[lambda f: ~f['params.model_name'].isna()]
 .drop_duplicates('params.model_name')
 [columns_to_show]  
 .head()
)

## Restauração de Experimentos

A partir da escolha de uma execução individual, é possível restaurar os elementos utilizados na experimentação para aplicá-los aos dados.

RUN_ID = '11bed0968e80465f8775b94e95accffc'

preprocessing_model, model, label_encoder = load_model_resources(RUN_ID)

Para validar o funcionamento da restauração do modelo, parte dos dados de treinamento são recuperados para uma avaliação.

columns_to_read = ['query', 'intent_class', 'intent_description']

frame = pd.read_csv(Path(settings.DATA_PATH).joinpath('interim', 'query_intent_training.csv'), usecols=columns_to_read)

display_side_by_side([frame.head(10)], ['Conjunto de Dados de Intenções de Busca'])

Tendo os dados, é possível refazer a preparação e fazer a predição da intenção de busca de cada elemento.

frame_slice = frame.sample(50)

# Processa os dados para inferência
features = preprocessing_model.predict(frame_slice)

# Realiza a inferência
frame_slice['pred'] = model.predict(features)

display_side_by_side([frame_slice, 
                      pd.DataFrame(features).describe().T.head(10)], 
                     ['Dados Recuperados e Predição', 
                      f'Features (10 de {features.shape[1]})'])

del frame_slice
