==> notebooks/05.0_Analise_de_Integracao_de_Servicos.ipynb <==
## Análise de Integração de Serviços para a Recomendação de Produtos

Este *notebook* tem como objetivo analisar o uso conjunto dos modelos criados para fornecer recomendações de produtos. A proposta é receber uma sentença de busca, identificar sua categoria (como se ela fosse um produto), reconhecer sua intenção e gerar recomendações de produtos usando essas informações.

## Bibliotecas e Funções

# General
import sys
import funcy as fp
from pathlib import Path
from typing import List

# Visualization / Presentation
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.core.display import HTML, display

# Data manipulation and computation
import numpy as np
import pandas as pd

# Embeddings and Similarity
from sklearn.metrics.pairwise import cosine_similarity
import fasttext

# Carregar, além de atualizar frequentemente, código personalizado disponível em ../src
%load_ext autoreload 
%autoreload 2
sys.path.append(str(Path.cwd().parent))
from src import settings
from src.utils.notebooks import display_side_by_side
from src.entities import Product
from src.pipeline.training_pipeline import compute_embeddings_frame
from src.pipeline.inference_pipeline import (load_model_resources,
                                             make_batch_predictions,
                                             make_supervised_intent_classification
                                            )

# Configurações para a exibição de conteúdo do Pandas e das bibliotecas gráficas
%matplotlib inline 
sns.set(rc={'figure.figsize':(25,10)})
pd.set_option('display.max_rows', None)
pd.set_option("display.max_columns", None)
pd.set_option('max_colwidth', 150)

## Recuperação dos Modelos

Considerando os experimentos feitos nos notebooks [Classificação de Categorias](03.0_Classificacao_de_Categorias.ipynb) e [Classificação de Intenções](04.3_Classificacao_de_Intencoes.ipynb), é possível fazer as tarefas de **classificação de categoria de produto** e de **classificação intenção de busca** usando as funções de inferência já adotadas para a [produtização](06_Consumo_Produtizacao.ipynb). Considerando duas buscas como ponto de partida, pode-se obter a categoria do produto a qual cada uma pertence  e a intenção da busca.

A classificação da busca como um produto é possível sem tratamentos adicionais considerando que o pré-processamento realiza imputação de valores ausentes para as características utilizadas.

queries = ["Roupinha de bebê",
           "Lembrancinha chá de bebê"]

products = [Product(title=query, price=None, concatenated_tags=None)
            for query in queries]

products_with_categories = make_batch_predictions(products)
query_intent = make_supervised_intent_classification(queries)

display_side_by_side([pd.DataFrame(products_with_categories)[['title', 'category']],
                      pd.DataFrame({'Intenções': query_intent})], 
                     ['Classificação de Produtos',
                      'Intenções de Busca'],
                    padding=50)

Com o resultado da utilização dos dois classificadores, têm-se a categoria e a intenção da busca. Uma primeira estratégia que pode ser explorada para combinar as informações dos classificadores é a forma de fazer a recuperação de dados para a priorizar as recomendações:
 - **Foco**: Se a pessoa tem uma intenção voltada para um produto específico, é interessante que os produtos exibidos fiquem restritos à categorias próximas. Mostrar uma variedade grande de tipos de produtos pode fazer a pessoa acreditar que os produtos semelhantes ao que ela queria já foram apresentados e não continuar explorando os resultados. Assim, pode-se restringir os resultados à categoria da busca.
 - **Exploração**: Caso a pessoa tenha o objetivo de explorar, é interessante apresentar maior diversificação para que ela possa ver diferentes opções para entender o que quer comprar ou possa refinar sua busca para então torná-la mais direcionada àquilo que entendeu atender sua necessidade. Com isso, é possível recuperar produtos de mais categorias.
 
Esse processo será explorado a seguir.

## Probabilidade de uma Busca ser de uma Categoria

Para conseguir identificar a classe predita para um produto (no caso de foco) ou múltiplas classes de produto que atendam a uma probabilidade mínima (para a intenção de exploração), é preciso personalizar o processo de classificação para se obter as probabilidades em vez da categoria final. O processo é realizado adiante.

CATEGORY_PROBABILITY_THRESHOLD = .25

#query = "Roupinha de bebê"
query = "Lembrancinha chá de bebê"

query_product = Product(title=query, price=None, concatenated_tags=query)

# Identificar a intenção de busca
query_intent = fp.first(make_supervised_intent_classification([query]))

# Recuperar recursos de pré-processamento e classificação
category_preprocessing_model, category_model, category_label_encoder_model = load_model_resources(settings.CATEGORY_CLASSIFICATION_RUN_ID)

# Pré-processar e classificar consulta como um produto
queries_frame = pd.DataFrame([query_product])
preprocessed_products = category_preprocessing_model.predict(queries_frame)
predictions = category_model.predict_proba(preprocessed_products)

# Obter categorias que atendam a um threshold de probabilidade (exploração) ou com a com maior probabilidade (foco)
categories = np.array(category_label_encoder_model.classes_)

if query_intent != 'Foco':
    selected_categories = set(categories[predictions[0] >= CATEGORY_PROBABILITY_THRESHOLD])
else:    
    selected_categories = set([categories[predictions[0].argmax()]])

# Criar dicionário com a probabilidade das categorias selecionadas
category_prob_dict = {category: prob 
                      for prob, category in zip(predictions[0], categories)
                      if category in selected_categories
                     }

display(HTML(f'<strong>Query</strong>: {query}'))
display(HTML(f'<strong>Query Intent</strong>: {query_intent}'))
display(HTML(f'<strong>Selected categories</strong>: {", ".join(selected_categories)}'))

## Pré-processamento para Recuperação

Com a consulta definida e as classes consideradas para a recuperação, de acordo com os critérios definidos para cada intenção, faz-se o carregamento e o pré-processamento dos dados.

columns_to_read = ['product_id', 'title', 'concatenated_tags', 'category', 'view_counts', 'order_counts']

frame = pd.read_csv(Path(settings.DATA_PATH).joinpath('interim', 'training.csv'), usecols=columns_to_read).drop_duplicates('product_id')
display_side_by_side([frame.head()], [f'Dados de Produtos (Registros: {len(frame)})'])

O pré-processamento envolve:
 - Fazer o filtro dos produtos pelas categorias consideradas;
 - Calcular os *embeddings* relacionados às variávies textuais (*title* e *concatenated tags*)
 - Atribuir a cada categoria de um produto a probabilidade de a busca pertencer a essa categoria.

embeddings_columns = ['title', 'concatenated_tags']
search_frame = (frame
                .loc[lambda f: f['category'].isin(selected_categories)]
                .assign(category_prob=lambda f: f['category'].map(category_prob_dict))
               )
search_frame = pd.concat([search_frame, compute_embeddings_frame(search_frame, embeddings_columns)], axis=1)


display_side_by_side([search_frame.head(2)], [f'Produtos Pré-Processados para Busca (Registros: {len(search_frame)})'])

## Recomendação de Produtos

Por fim, são calculados os atributos de similaridade e a computação de relevância para ordenar os resultados:
 - Gerar o *embedding* da busca;
 - Calcular a similaridade entre *embedding* da busca e os *embeddings* do título e das tags concatenadas;
 - Calcular a eficácia de venda de cada produto, considerando *compras*/*visualizações*, limitando o valor a no máximo 1 (caso o número de compras seja maior do que o de visualizações);
 - Calcular a pontuação final do produto considerando a similaridade dos embeddings, probabilidade de a busca pertencer à mesma categoria do produto e o número de compras por visualizações; 
 - Ordenar os produtos pela pontuação final em ordem decrescente e usar a eficácia de venda como critério de desempate caso dois produtos tenham a mesma pontuação.
 
Para cada um dos 3 elementos de similaridade (2 *embeddings* e probabilidade da categoria), além do número de vendas por visualizações, possui um peso atribuído para indicar sua contribuição para a pontuação final. Nesta versão, esses pesos são atribuídos manualmente, mas é possível construir um modelo para identificar qual o melhor peso de cada fator. Um modelo poderia ir além, podendo utilizar os *embeddings* da busca e dos elementos dos produtos para achar relações mais sofisticadas entre os números e que podem ser mais relevantes para determinar a interação com um produto.

def compute_embedding_columns_similarity(base_embeddings_frame: pd.DataFrame,
                                         reference_embedding: np.array,
                                         columns: List[str]) -> pd.DataFrame:
    """Adds, for each column, the similarity between its elements and the reference embedding"""

    similarity_frame = base_embeddings_frame[[]].copy()
    reference_embedding = np.expand_dims(reference_embedding, axis=0)

    for column in columns:
        column_embeddings = np.stack(base_embeddings_frame[f'{column}_embedding'].to_numpy(), axis=0)
        similarity_frame[f'{column}_similarity'] = np.concatenate(cosine_similarity(reference_embedding, column_embeddings), axis=0)

    return similarity_frame


def search_products(base_frame: pd.DataFrame,
                    query: str,
                    items_to_retrieve: int = 10
                   ) -> pd.DataFrame:
    
    TITLE_WEIGHT = 1.0
    TAGS_WEIGHT = 0.75
    CATEGORY_WEIGHT = 0.5
    ORDER_PER_VIEW_WEIGHT = 0.25

    ft_model = fasttext.load_model(str(Path(settings.MODELS_PATH).joinpath(settings.EMBEDDINGS_MODEL)))
    query_embedding = ft_model.get_sentence_vector(query)

    search_frame = base_frame.copy()
    search_frame = pd.concat([search_frame, 
                             compute_embedding_columns_similarity(search_frame, query_embedding, ['title', 'concatenated_tags'])]
                            , axis=1)
    
    columns_to_drop = [item for item in search_frame.columns if item.endswith('_embedding')]    

    return (search_frame
            .fillna({'order_counts': 0,
                     'view_counts': 0})
            .assign(orders_per_views=lambda f: ((f['order_counts']+1)/(f['view_counts']+1)).apply(lambda x: min(1.0, x)))
            .assign(score=lambda f: 
                    (f['title_similarity'] * TITLE_WEIGHT) + 
                    (f['concatenated_tags_similarity'] * TAGS_WEIGHT) + 
                    (f['category_prob'] * CATEGORY_WEIGHT) + 
                    (f['orders_per_views'] * ORDER_PER_VIEW_WEIGHT)
                   )
            .sort_values(by='score', ascending=False)
            .head(items_to_retrieve)
            .drop(columns=columns_to_drop, inplace=False)
           )    
    

display_side_by_side([search_products(search_frame, query, 20)],
                     [f'Resultados para : \'{query}\''])
