==> notebooks/06.0_Consumo_Produtizacao.ipynb <==
# Integração e Uso dos Sistemas "em Produção"

O objetivo deste *notebook* é documentar a utilização dos sistemas criados por meio de uma *Command Line Interface* criada no projeto. Com isso, tem-se a simulação de como o sistema seria fornecido para uso em ambiente de produção ambiente de produção.

## Bibliotecas e Configurações

Dada a proposta de registrar o uso de uma CLI, o uso dela é auto-contido e minimiza a necessidade de importação de módulos do Python. A única exceção é o uso de uma biblioteca para fazer a representação de dicionários como JSON. Essa exceção serve apenas como forma de a serialização do conteúdo e enviá-lo para a CLI.

import json
from IPython.core.display import HTML, display

PYTHON_PATH = 'cd ..;../virtual_envs/teste_HT/bin/python '

COLOR_RESULT = '#5499c7'
COLOR_COMMAND = '#bb8fce'

TEXT_RESULT = 'Resultado'
TEXT_COMMAND = 'Comando'

BASE_STYLE = '''
    width:100%;
    padding:10px 10px 10px 10px;
    text-align:left;
    background-color:{color};
    font-family: Courier New,Courier,Lucida Sans Typewriter,Lucida Typewriter,monospace;
'''

def format_output(content: str, title: str, color: str='#ccd1d1') -> None:
    style = BASE_STYLE.format(color=color)
    display(HTML(f'''
        <span style="color:{color};font-weight:bold">{title}</span>
        <div style="{style}">{content}</div>'''))

## Uso dos Módulos

### Classificação de Categorias

Os exemplos foram retirados dos primeiros itens que foram apresentados no [site](https://elo7.com) a partir da seleção da categoria. Para evitar um conteúdo excessivamente longo, apenas as primeiras *tags* dos produtos foram utilizadas.

category_classification_tests = [
    {"title":"CARIMBO LOGOMARCA", 
     "concatenated_tags": "carimbo escola scrapbook decoupage carimbo para scrapbook carimbo incentivo", 
     "price": 30},

    {"title":"Caneca Personalizada Madrinha/Padrinho", 
     "concatenated_tags": "caneca coruja caneca brinde caneca aniversário caneca personagens caneca infantil caneca dia dos namorados", 
     "price": 9.70},

    {"title":"Pelego Fake Branco", 
     "concatenated_tags": "pelego pelego fake pelucia cadeira rose rosa milenium decor decoracao home office escritorio", 
     "price": 39.90},

    {"title":"Kit Pulseiras Now United", 
     "concatenated_tags": "bijuterias bijuterias folheadas bijuterias delicadas bijus pulseira pulseira com pingentes pulseira delicada ", 
     "price": 19.70},

    {"title":"Wrap sling de algodão", 
     "concatenated_tags": "comprar wrap sling carregador de bebê canguru 100% algodão", 
     "price": 58.90},

    {"title":"Sapateira Multiuso de Madeira P/ Porta Sala Quarto", 
     "concatenated_tags": "madeira sapatos sapateira porta sapatos tenis sapateira de sapateira de madeira multiuso de madeira pinus estante", 
     "price": 64.99},
]

for ix, test_case in enumerate(category_classification_tests):
    
    display(HTML(f'<br /><h4>Test Case {ix + 1}</h4>'))
    
    command = f'python cli/teste_ds.py --category \'{json.dumps(test_case)}\''
    format_output(command, TEXT_COMMAND, COLOR_COMMAND)

    # Send the test case content as a parameter for the CLI and store the result
    cli_outputs = !{PYTHON_PATH} cli/teste_ds.py --category {"\'" + json.dumps(test_case) + "\'"}

    # Show only the last printed output to suppress system messages and show a clean reponse
    format_output(cli_outputs[-1], TEXT_RESULT, COLOR_RESULT)

### Intenção de Busca

Para as intenções de busca, apenas uma sentença simples pode ser definida como entrada.  Para os exemplos, foram consideradas buscas que permitissem obter produtos específicos disponíveis no [Elo7](http://elo7.com.br) ou que tivessem uma tendência a trazer resultados de um escopo amplo.

query_intent_classification_tests = [
    'máscara de tecido tripla camada padrão oms',
    'presente de aniversário',
    'caneca personalizada de star wars',
    'lembrancinha',
    'peças de decoração',
    'papel fotográfico glossy folha a4',
]

for ix, test_case in enumerate(query_intent_classification_tests):

    display(HTML(f'<br /><h4>Test Case {ix + 1}</h4>'))

    command = f'python cli/teste_ds.py --intent \'{test_case}\'  \n'
    format_output(command, TEXT_COMMAND, COLOR_COMMAND)

    # Send the test case content as a parameter for the CLI and store the result
    cli_outputs = !{PYTHON_PATH} cli/teste_ds.py --intent {"\'" + test_case + "\'"}

    # Show only the last printed output to suppress system messages and show a clean reponse
    format_output(cli_outputs[-1], TEXT_RESULT, COLOR_RESULT)    

## Recommendação de Produtos (Colaboração de Sistemas)

A recomendação de produtos envolve a aplicação dos modelos de classificação de produtos em categorias, classificação de itenções e recomendação, com a filtragem e ordenação a partir do conteúdo. Apesar do uso de todos esses recursos criados, apenas uma sentença de busca é fornecida como entrada. Assim como ocorreu anteriormente, tentou-se definir ao menos um exemplo de buscas com escopo restrito e amplo, para avaliar os resultados.

queries_tests = [
    'presente de casamento',
    'lembrancinha de chá de bebê',
    'papel fotográfico glossy folha A4'
]

for ix, test_case in enumerate(queries_tests):
    display(HTML(f'<br /><h4>Test Case {ix + 1}</h4>'))

    command = f'python cli/teste_ds.py --recommendation \'{test_case}\'  \n'
    format_output(command, TEXT_COMMAND, COLOR_COMMAND)

    # Send the test case content as a parameter for the CLI and store the result
    cli_outputs = !{PYTHON_PATH} cli/teste_ds.py --recommendation {"\'" + test_case + "\'"}

    # Show only the last printed output to suppress system messages and show a clean reponse
    formatted_output = '<br />'.join(cli_outputs[-10:])
    format_output(formatted_output, TEXT_RESULT, COLOR_RESULT)
